update the battleship game
Added the ability to change world size via the board_size variable. Updated the code to react to the board_size changes. Added the base code to generate ship seeds based on size. Next step is to add those generated ships to a dictionary in order to check for hits.  Ships will consist of [Type(String, [X](int), [Y](int)]. Ships will be generated incrementally from the seed in an x or y direction based on a "coin flip".